api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow>=10.3.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
ollama>=0.1.0
gunicorn==21.2.0
